sample machine learning application/deps/data_ingress_rest_requirements.txt
kafka-python
tornado